Fix coverity issues in execHHashagg.c
CID 167715: Incorrect Expression (UNINTENDED_INTEGER_DIVISION)
Covertiy reported that a division operation of two integers is then
converted to a double. This would cause any fractional part of the
quotient to be lost. This is fixed by casting the divisor to a double
before the division.

CID 167714: Dereference before null check (REVERSE_INULL)
Coverity reported that a null check is unnecessary because the pointer
spill_file->file_info is dereferenced before this line in every code
path. This issue is fixed by removing the null pointer check